Finally consolidate both create functions (one extracted from Gruntfile, the other from webpack.js)
